split information in sections
