+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/profile/body/application.tsx
@@ -2,10 +2,13 @@ import * as React from 'react';
 import {connect, Dispatch} from 'react-redux';
 import ApplicationPanel from '~/components/general/application-panel';
 import { StateType } from 'reducers';
-import ProfileInfoAndSettings from './application/profile-info-and-settings';
-import ProfilePicture from './application/profile-picture';
+import ChatSettings from "./application/chat-settings";
+import ContactInformation from "./application/contact-information";
+import GeneralInformation from "./application/general-information";
+import Security from "./application/security";
+import VacationSettings from "./application/vacation-settings";
+import WorkList from "./application/work-list";
 import { i18nType } from '~/reducers/base/i18n';
-
 
 interface ProfileApplicationProps {
   i18n: i18nType,
@@ -19,9 +22,18 @@ class ProfileApplication extends React.Component<ProfileApplicationProps, Profil
   render(){
     let title = this.props.i18n.text.get('plugin.profile.profile')
     return (<div className="application-panel-wrapper">
-      <ApplicationPanel modifier="profile" title={title} asideBefore={this.props.aside} disableStickyScrolling>
-        <ProfilePicture/>
-        <ProfileInfoAndSettings/>
+      <ApplicationPanel
+        modifier="profile"
+        title={title}
+        asideBefore={this.props.aside}
+        disableStickyScrolling
+      >
+        <ChatSettings />
+        <ContactInformation />
+        <GeneralInformation />
+        <Security />
+        <VacationSettings />
+        <WorkList />
       </ApplicationPanel>
     </div>);
   }
