+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/profile/body/application/general-information.tsx
@@ -1,0 +1,76 @@
+import * as React from "react";
+import { connect } from "react-redux";
+import { StateType } from "~/reducers";
+import { i18nType } from "~/reducers/base/i18n";
+import { StatusType } from "~/reducers/base/status";
+import { ProfileType } from "~/reducers/main-function/profile";
+import ProfilePicture from "./components/profile-picture";
+import ProfileProperty from "./components/profile-property";
+import moment from '~/lib/moment';
+
+interface IGeneralInformationProps {
+  i18n: i18nType,
+  profile: ProfileType;
+  status: StatusType;
+}
+
+interface IGeneralInformationState {
+  profileVacationStart: any;
+  profileVacationEnd: any;
+  phoneNumber: string,
+  chatVisibility: string,
+  chatNickname: string,
+  vacationAutoReply: string,
+  vacationAutoReplySubject: string,
+  vacationAutoReplyMsg: string,
+}
+
+class GeneralInformation extends React.Component<IGeneralInformationProps, IGeneralInformationState> {
+  constructor(props: IGeneralInformationProps) {
+    super(props);
+  }
+
+  public render() {
+    if (this.props.profile.location !== "general") {
+      return null;
+    }
+
+    const studyTimeEndValues = [];
+    if (this.props.status.profile.studyTimeEnd){
+      studyTimeEndValues.push(this.props.i18n.time.format(moment(this.props.status.profile.studyTimeEnd, "ddd MMM DD hh:mm:ss ZZ YYYY").toDate()));
+      if (this.props.status.profile.studyTimeLeftStr){
+        studyTimeEndValues.push(this.props.status.profile.studyTimeLeftStr);
+      }
+    }
+
+    return (
+      <div className="profile-element">
+        <ProfilePicture />
+
+        <h2 className="profile-element__title">{this.props.status.profile.displayName}</h2>
+
+        <ProfileProperty i18n={this.props.i18n} condition={!!this.props.status.profile.studyStartDate} label="plugin.profile.studyStartDateLabel"
+          value={this.props.i18n.time.format(moment(this.props.status.profile.studyStartDate, "ddd MMM DD hh:mm:ss ZZ YYYY").toDate())} />
+        <ProfileProperty i18n={this.props.i18n} condition={!!this.props.status.profile.studyTimeEnd} label="plugin.profile.studyTimeEndLabel"
+          value={studyTimeEndValues} />
+      </div>
+    );
+  }
+}
+
+function mapStateToProps(state: StateType) {
+  return {
+    i18n: state.i18n,
+    profile: state.profile,
+    status: state.status,
+  }
+};
+
+function mapDispatchToProps(dispatch: React.Dispatch<any>) {
+  return {};
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(GeneralInformation);
